fix instance restarting fight.
When a Lua coroutine (thread) finds that another thread is starting an
instance, it waits until the other either succeeds or fails.  In case of
failure, the original thread fails the request, and the second one
tries again.  This retry was wrongly being stored in the old intance
info table which is removed by the failing thread, and never stored
back in the general table by the retrying thread.  This fix ensures
that on the failing case the second thread doesn't hold the old info,
but creates a new one.